2003-08-07  Al Riddoch  <ajr@ecs.soton.ac.uk>
	* visual/TerrainRenderer.cpp, visual/TerrainRenderer.h:
	  Use short for array indexes.

+++ visual/TerrainRenderer.h
@@ -16,7 +16,7 @@ class TerrainRenderer : public EntityRenderer
   protected:
     Mercator::Terrain m_terrain;
     int m_numLineIndeces;
-    unsigned int * const m_lineIndeces;
+    unsigned short * const m_lineIndeces;
     float * const m_texCoords;
     GLuint m_textures[8]; //, m_texture2;
     bool m_haveTerrain;
